user

+++ bot.py
@@ -28,24 +28,27 @@ dp = Dispatcher(bot, storage=storage)
 ADMINS = ['blacktyg3r']
 
 COMMANDS = ['help', 'add', 'remove', 'delete', 'del',
-            'done', 'tick', 'liked', 'REMOVE_ALL']
+            'done', 'tick', 'like', 'REMOVE_ALL']
 short_time = 20
 last_show_msg = 0
+last_help_msg = 0
 
 
 @dp.message_handler(commands=['help'])
 async def help_command(message: types.Message):
-    reply = Response('🧙🏼‍♂', "WednesdayWizard! 💬")
+    reply = Response('🧙🏼‍♂', "<b>@WednesdayWizardBot</b>! 💬")
     lines = help_msg()
-    lines = [
-        reply.separator
-        ]
+    lines.append(reply.separator)
     reply.lines = lines
     reply.add_lines()
     send = await message.reply(f'{reply.print()}', parse_mode=ParseMode.HTML, reply=False)
-    await asyncio.sleep(short_time * 2)
-    await send.delete()
+    global last_help_msg
+    if last_help_msg:
+        await bot.delete_message(message.chat.id, last_help_msg)
+    last_help_msg = send.message_id
+    await asyncio.sleep(short_time)
     await message.delete()
+    await send.delete()
 
 
 @dp.message_handler(commands=['add'])
@@ -73,10 +76,15 @@ async def add_command(message: types.Message):
 @dp.message_handler(commands=['show'])
 async def show_command(message: types.Message):
     if message.is_command():
-        length = message.get_args() or 5
+        if message.get_args() != '':
+            length = int(message.get_args())
+        else:
+             length = 5
         reply = Response('🧙🏼‍♂', "MAGIC LINK LIST 🖇")
-        lines = [f"ID: <b>{i}</b> | {link.time} | {icon_number(link.votes)} 👍\n{link.url}"
-                 for i, link in enumerate(db.get()['links'])][:length]
+        lines = [f"LINK_ID: <b>{i}</b> | {link.time} | {icon_number(link.votes)} 👍\n{link.url}"
+                 for i, link in enumerate(db.get()['links'])]
+        lines.reverse()
+        lines = lines[:length]
         reply.lines = lines
         reply.add_lines()
         print(db.get())
@@ -84,41 +92,34 @@ async def show_command(message: types.Message):
                                    disable_notification=True, disable_web_page_preview=True)
         global last_show_msg
         if last_show_msg:
-            try:
-                await bot.delete_message(message.chat.id, last_show_msg)
-            except:
-                pass
-
+            await bot.delete_message(message.chat.id, last_show_msg)
         last_show_msg = send.message_id
         await asyncio.sleep(short_time)
         await message.delete()
-        await asyncio.sleep(short_time)
-        try:
-            await send.delete()
-        except:
-            pass
+        await send.delete()
 
 
-@dp.message_handler(commands=['done', 'liked', 'tick'])
+@dp.message_handler(commands=['done', 'like', 'tick'])
 async def done_command(message: types.Message):
     if message.is_command():
         ids = message.get_args().split(' ')
         user = message.from_user.username
         for id in ids:
+            print(id)
             link = db.get_links()[int(id)]
             if user not in link.voters:  # and user != link['author']:
                 link.voters.append(user)
                 link.votes += 1
                 db.update_link(link)
-                reply = f"🧙🏼‍♂ @{user} IS BLESSING LINK ID: <b>{', '.join(ids)}</b> 🎊"
+                reply = f"🧙🏼‍♂ @{user} IS BLESSING LINK_ID: <b>{', '.join(ids)}</b> 🎊"
             else:
-                reply = f"🧙🏼‍♂ @{user}, LINK ID: <b>{', '.join(ids)}</b> ALREADY BLESSED!"
+                reply = f"🧙🏼‍♂ @{user}, LINK_ID: <b>{', '.join(ids)}</b> ALREADY BLESSED!"
 
-            send = await message.reply(f'{reply}', parse_mode=ParseMode.HTML, reply=False,
-                                       disable_notification=True, disable_web_page_preview=True)
-            await asyncio.sleep(int(short_time / 2))
-            await message.delete()
-            await send.delete()
+        send = await message.reply(f'{reply}', parse_mode=ParseMode.HTML, reply=False,
+                                   disable_notification=True, disable_web_page_preview=True)
+        await asyncio.sleep(int(short_time / 2))
+        await message.delete()
+        await send.delete()
 
 
 @dp.message_handler(commands=['delete', 'remove', 'del'])
@@ -131,9 +132,9 @@ async def delete_command(message: types.Message):
             if user == link.author or user in ADMINS:
                 db.remove_link(link)
                 print('REMOVED:', link)
-                reply = f"🧙🏼‍♂ @{user} MADE LINK ID: <b>{', '.join(ids)}</b> | DISAPPEAR! "
+                reply = f"🧙🏼‍♂ @{user} MADE LINK_ID: <b>{', '.join(ids)}</b> | DISAPPEAR! "
             else:
-                reply = f"🧙🏼‍♂ @{user}, LINK ID: <b>{', '.join(ids)}</b> IS NOT YOURS!"
+                reply = f"🧙🏼‍♂ @{user}, LINK_ID: <b>{', '.join(ids)}</b> IS NOT YOURS!"
 
         send = await message.reply(f'{reply}', parse_mode=ParseMode.HTML, reply=False,
                                    disable_notification=True, disable_web_page_preview=True)
@@ -164,7 +165,8 @@ async def delete_all_command(message: types.Message):
 async def delete_all_command(message: types.Message):
     if message.is_command():
         if message.get_command() not in COMMANDS:
-            await message.delete()
+            if message.get_args() != "":
+                await message.delete()
 
 
 # START BOT
